requirements.txt
fastapi==0.115.6
uvicorn==0.32.1
requests==2.32.5
python-dotenv==1.0.1
aiofiles==24.1.0
azure-ai-projects>=1.0.0b9
azure-identity>=1.25.0
azure-search-documents>=11.6.0
